logout button appears while logged in, making new front end file to try adding vehicle info form add form
