+++ models/vehicle.js
@@ -7,7 +7,7 @@ Vehicle.init(
     {
         user_id: {
             type: DataTypes.INTEGER,
-            allowNull: false,
+            allowNull: true,
             references: {
                 model: 'user',
                 key: 'id',
