+++ views/homepage.handlebars
@@ -1,13 +1,5 @@
 {{#if logged_in}}
 <div class="sidenav">
-  <ul class="nav">
-    <li class="nav-item">
-      <button id="homeBtn">Home</button>
-    </li>
-    <li class="nav-item">
-      <button id="logoutBtn">Logout</button>
-    </li>
-  </ul>
   <a href="/vinForm">Add new vehicle</a>
   <section class="container">
     <section class="flex-row justify-center">
