+++ views/layouts/main.handlebars
@@ -17,21 +17,12 @@
 
 <main>
   {{#if logged_in}}
-  <div class="sidenav">
-  <ul class="nav">
-    <li class="nav-item">
-      <button id="homeBtn">Home</button>
-    </li>
-    <li class="nav-item">
-      <button id="logoutBtn">Logout</button>
-    </li>
-  </ul>
-  </div>
+  <button id="logoutBtn">Logout</button>
   {{/if}}
   {{{body}}}
 
 </main>
-
+<script src="/js/main.js"></script>
 </body>
 
 </html>
