+++ views/vinForm.handlebars
@@ -1,16 +1,3 @@
-{{#if logged_in}}
-  <div class="sidenav">
-  <ul class="nav">
-    <li class="nav-item">
-      <button id="homeBtn">Home</button>
-    </li>
-    <li class="nav-item">
-      <button id="logoutBtn">Logout</button>
-    </li>
-  </ul>
-  </div>
-  {{/if}}
-
 <form>
   <div class="mb-3">
     <label for="vin" class="form-label">*Vin: </label>
@@ -46,3 +33,5 @@
   </div>
   <button type="submit" class="btn btn-primary" id="addByYmm">Add</button>
 </form>
+
+{{!-- make js file for this page/form --}}
